Moved compare_files() to cli_helpers.py and updated related tests.
