+++ src/uwtools/utils/cli_helpers.py
@@ -67,3 +67,22 @@ def setup_logging(user_args, log_name=None): #pylint: disable=unused-variable
         log.propagate = False
 
     return log
+
+def compare_files(expected, actual):
+    '''Compare the content of two files.  Doing this over filecmp.cmp since 
+    we may not be able to handle end-of-file character differences with it.
+    Prints the contents of two compared files to std out if they do not match.'''
+
+    with open(expected, 'r', encoding='utf-8') as expected_file:
+        expected_content = expected_file.read().rstrip('\n')
+    with open(actual, 'r', encoding='utf-8') as actual_file:
+        actual_content = actual_file.read().rstrip('\n')
+
+    if expected_content != actual_content:
+        print('The expected file looks like:')
+        print(expected_content)
+        print('*' * 80)
+        print('The rendered file looks like:')
+        print(actual_content)
+        return False
+    return True
